Tratamento para respostas nulas da Query

+++ src/main/java/br/com/joaosarmento/listadepartidasapi/DTOs/RetrospectivaDTO.java
@@ -1,9 +1,9 @@
 package br.com.joaosarmento.listadepartidasapi.DTOs;
 
 public interface RetrospectivaDTO {
-     Long getVitorias();
-     Long getDerrotas();
-     Long getEmpates();
-     int getGolsPro();
-     int getGolsContra();
+     Integer getVitorias();
+     Integer getDerrotas();
+     Integer getEmpates();
+     Integer getGolsPro();
+     Integer getGolsContra();
 }
